Move generic lane-size tasks into celery/tasks/lanes.py
The lane-size machinery does not belong in the custom_lists module. Split
it out into a dedicated celery/tasks/lanes.py:

- update_lane_size (per-lane worker)
- finalize_lane_size_update (chord callback / cache invalidation)
- update_independent_lane_sizes (sweep of non-custom-list lanes)

custom_lists.py keeps the custom-list-specific pieces and imports the two
primitives back from lanes.py (custom_lists -> lanes). The lane-classification
query is made public (custom_list_lane_ids_query) since lanes.py now consumes
it. The two lanes -> custom_lists needs (the classification query and the
sweep-lock identity used by the finalize release branch) are function-local
imports to keep module load acyclic.

Beat schedule and UpdateIndependentLaneSizesScript now reference
lanes.update_independent_lane_sizes. Tests for the moved tasks relocate to
tests/manager/celery/tasks/test_lanes.py. No behaviour change.
